Refactor import statement for Simulation class and add new simulation module
